# fetch: replace HeadersList Array inheritance to Map