--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -91,61 +91,33 @@ function fill (headers, object) {
   }
 }
 
-// TODO: Composition over inheritence? Or helper methods?
-class HeadersList extends Array {
+class HeadersList extends Map {
   append (name, value) {
     const normalizedName = normalizeAndValidateHeaderName(name)
     const normalizedValue = normalizeAndValidateHeaderValue(name, value)
 
-    const index = binarySearch(this, normalizedName)
+    const exists = super.get(normalizedName)
 
-    if (this[index] === normalizedName) {
-      this[index + 1] += `, ${normalizedValue}`
+    if (exists) {
+      this.set(normalizedName, `${exists}, ${normalizedValue}`)
     } else {
-      this.splice(index, 0, normalizedName, normalizedValue)
+      this.set(normalizedName, `${normalizedValue}`)
     }
   }
 
   delete (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-
-    const index = binarySearch(this, normalizedName)
-
-    if (this[index] === normalizedName) {
-      this.splice(index, 2)
-    }
+    return super.delete(normalizedName)
   }
 
   get (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-
-    const index = binarySearch(this, normalizedName)
-
-    if (this[index] === normalizedName) {
-      return this[index + 1]
-    }
-
-    return null
+    return super.get(normalizedName) || null
   }
 
   has (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-
-    const index = binarySearch(this, normalizedName)
-
-    return this[index] === normalizedName
-  }
-
-  set (name, value) {
-    const normalizedName = normalizeAndValidateHeaderName(name)
-    const normalizedValue = normalizeAndValidateHeaderValue(name, value)
-
-    const index = binarySearch(this, normalizedName)
-    if (this[index] === normalizedName) {
-      this[index + 1] = normalizedValue
-    } else {
-      this.splice(index, 0, normalizedName, normalizedValue)
-    }
+    return super.has(normalizedName)
   }
 }
 
@@ -160,17 +132,24 @@ class Headers {
         "Failed to construct 'Headers': The provided value is not of type '(record<ByteString, ByteString> or sequence<sequence<ByteString>>"
       )
     }
-    const init = args.length >= 1 ? args[0] ?? {} : {}
-
-    this[kHeadersList] = new HeadersList()
-
+    let init = {}
     // The new Headers(init) constructor steps are:
 
     // 1. Set this’s guard to "none".
     this[kGuard] = 'none'
 
     // 2. If init is given, then fill this with init.
-    fill(this, init)
+    // TODO(rafaelgss): it replaces fill
+    if (args.length >= 1) {
+      init = args[0] ?? {}
+      if (init[Symbol.iterator]) {
+        this[kHeadersList] = new HeadersList(init)
+      } else if (typeof init === 'object') {
+        this[kHeadersList] = new HeadersList(Object.entries(init))
+      } else {
+        throw new TypeError()
+      }
+    }
   }
 
   get [Symbol.toStringTag] () {
@@ -309,23 +288,25 @@ class Headers {
   }
 
   * keys () {
-    const clone = this[kHeadersList].slice()
-    for (let index = 0; index < clone.length; index += 2) {
-      yield clone[index]
+    console.log('called')
+    const listFreezed = new HeadersList(this[kHeadersList])
+    for (const key of listFreezed.keys()) {
+      yield key
     }
   }
 
   * values () {
-    const clone = this[kHeadersList].slice()
-    for (let index = 1; index < clone.length; index += 2) {
-      yield clone[index]
+    const listFreezed = new HeadersList(this[kHeadersList])
+    for (const val of listFreezed.values()) {
+      yield val
     }
   }
 
   * entries () {
-    const clone = this[kHeadersList].slice()
-    for (let index = 0; index < clone.length; index += 2) {
-      yield [clone[index], clone[index + 1]]
+    console.log('called')
+    const listFreezed = new HeadersList(this[kHeadersList])
+    for (const [key, val] of listFreezed.entries()) {
+      yield [key, val]
     }
   }
 
@@ -346,15 +327,7 @@ class Headers {
     const callback = args[0]
     const thisArg = args[1]
 
-    const clone = this[kHeadersList].slice()
-    for (let index = 0; index < clone.length; index += 2) {
-      callback.call(
-        thisArg,
-        clone[index + 1],
-        clone[index],
-        this
-      )
-    }
+    this[kHeadersList].forEach(callback.bind(thisArg))
   }
 
   [Symbol.for('nodejs.util.inspect.custom')] () {