--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -149,6 +149,8 @@ class Headers {
       } else {
         throw new TypeError()
       }
+    } else {
+      this[kHeadersList] = new HeadersList()
     }
   }