--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -112,7 +112,7 @@ class HeadersList extends Map {
 
   get (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-    return super.get(normalizedName) || null
+    return super.get(normalizedName) ?? null
   }
 
   has (name) {
@@ -121,6 +121,7 @@ class HeadersList extends Map {
   }
 }
 
+// https://fetch.spec.whatwg.org/#headers-class
 class Headers {
   constructor (...args) {
     if (
@@ -132,26 +133,16 @@ class Headers {
         "Failed to construct 'Headers': The provided value is not of type '(record<ByteString, ByteString> or sequence<sequence<ByteString>>"
       )
     }
-    let init = {}
+    const init = args.length >= 1 ? args[0] ?? {} : {}
+    this[kHeadersList] = new HeadersList()
+
     // The new Headers(init) constructor steps are:
 
     // 1. Set this’s guard to "none".
     this[kGuard] = 'none'
 
     // 2. If init is given, then fill this with init.
-    // TODO(rafaelgss): it replaces fill
-    if (args.length >= 1) {
-      init = args[0] ?? {}
-      if (init[Symbol.iterator]) {
-        this[kHeadersList] = new HeadersList(init)
-      } else if (typeof init === 'object') {
-        this[kHeadersList] = new HeadersList(Object.entries(init))
-      } else {
-        throw new TypeError()
-      }
-    } else {
-      this[kHeadersList] = new HeadersList()
-    }
+    fill(this, init)
   }
 
   get [Symbol.toStringTag] () {
@@ -289,25 +280,27 @@ class Headers {
     return this[kHeadersList].set(String(args[0]), String(args[1]))
   }
 
+  // The value pairs to iterate over are the return value of running sort
+  // and combine with this’s header list.
   * keys () {
-    console.log('called')
-    const listFreezed = new HeadersList(this[kHeadersList])
+    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
     for (const key of listFreezed.keys()) {
       yield key
     }
   }
 
   * values () {
-    const listFreezed = new HeadersList(this[kHeadersList])
+    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
     for (const val of listFreezed.values()) {
       yield val
     }
   }
 
+  // The value pairs to iterate over are the return value of running sort
+  // and combine with this’s header list.
   * entries () {
-    console.log('called')
-    const listFreezed = new HeadersList(this[kHeadersList])
-    for (const [key, val] of listFreezed.entries()) {
+    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
+    for (const [key, val] of listFreezed) {
       yield [key, val]
     }
   }
@@ -329,7 +322,8 @@ class Headers {
     const callback = args[0]
     const thisArg = args[1]
 
-    this[kHeadersList].forEach(callback.bind(thisArg))
+    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
+    listFreezed.forEach(callback.bind(thisArg))
   }
 
   [Symbol.for('nodejs.util.inspect.custom')] () {