--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -11,22 +11,8 @@ const {
   forbiddenResponseHeaderNames
 } = require('./constants')
 
-function binarySearch (arr, val) {
-  let low = 0
-  let high = Math.floor(arr.length / 2)
-
-  while (high > low) {
-    const mid = (high + low) >>> 1
-
-    if (val.localeCompare(arr[mid * 2]) > 0) {
-      low = mid + 1
-    } else {
-      high = mid
-    }
-  }
-
-  return low * 2
-}
+const kHeadersMap = Symbol('headers map')
+const kSorted = Symbol('headers map sorted')
 
 function normalizeAndValidateHeaderName (name) {
   if (name === undefined) {
@@ -91,33 +77,64 @@ function fill (headers, object) {
   }
 }
 
-class HeadersList extends Map {
+class HeadersList {
+  constructor(init) {
+    if (init instanceof HeadersList) {
+      this[kHeadersMap] = new Map(init[kHeadersMap])
+      this[kSorted] = init[kSorted]
+    } else {
+      this[kHeadersMap] = new Map(init)
+      this[kSorted] = false
+    }
+  }
+
   append (name, value) {
     const normalizedName = normalizeAndValidateHeaderName(name)
     const normalizedValue = normalizeAndValidateHeaderValue(name, value)
 
-    const exists = super.get(normalizedName)
+    const exists = this[kHeadersMap].get(normalizedName)
 
     if (exists) {
-      this.set(normalizedName, `${exists}, ${normalizedValue}`)
+      this[kHeadersMap].set(normalizedName, `${exists}, ${normalizedValue}`)
     } else {
-      this.set(normalizedName, `${normalizedValue}`)
+      this[kHeadersMap].set(normalizedName, `${normalizedValue}`)
+      this[kSorted] = false
     }
   }
 
   delete (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-    return super.delete(normalizedName)
+    return this[kHeadersMap].delete(normalizedName)
   }
 
   get (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-    return super.get(normalizedName) ?? null
+    return this[kHeadersMap].get(normalizedName) ?? null
   }
 
   has (name) {
     const normalizedName = normalizeAndValidateHeaderName(name)
-    return super.has(normalizedName)
+    return this[kHeadersMap].has(normalizedName)
+  }
+
+  sortMap() {
+    this[kHeadersMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
+    this[kSorted] = true
+  }
+
+  // bypass normalization
+  set (key, val) {
+    return this[kHeadersMap].set(key, val)
+  }
+
+  *[Symbol.iterator]() {
+    if (!this[kSorted]) {
+      this.sortMap()
+    }
+
+    for (const [key, val] of this[kHeadersMap].entries()) {
+      yield [key, val]
+    }
   }
 }
 
@@ -282,25 +299,31 @@ class Headers {
 
   // The value pairs to iterate over are the return value of running sort
   // and combine with this’s header list.
+  _getHeadersListIterable() {
+    if (!this[kHeadersList][kSorted]) {
+      this[kHeadersList].sortMap()
+    }
+
+    return new HeadersList(this[kHeadersList])
+  }
+
   * keys () {
-    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
-    for (const key of listFreezed.keys()) {
+    const listFreezed = this._getHeadersListIterable()
+    for (const key of listFreezed[kHeadersMap].keys()) {
       yield key
     }
   }
 
   * values () {
-    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
-    for (const val of listFreezed.values()) {
+    const listFreezed = this._getHeadersListIterable()
+    for (const val of listFreezed[kHeadersMap].values()) {
       yield val
     }
   }
 
-  // The value pairs to iterate over are the return value of running sort
-  // and combine with this’s header list.
   * entries () {
-    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
-    for (const [key, val] of listFreezed) {
+    const listFreezed = this._getHeadersListIterable()
+    for (const [key, val] of listFreezed[kHeadersMap]) {
       yield [key, val]
     }
   }
@@ -322,8 +345,10 @@ class Headers {
     const callback = args[0]
     const thisArg = args[1]
 
-    const listFreezed = new HeadersList([...this[kHeadersList]].sort())
-    listFreezed.forEach(callback.bind(thisArg))
+    const listFreezed = this._getHeadersListIterable()
+    listFreezed[kHeadersMap].forEach((value, index) => {
+      callback.apply(thisArg, [value, index, this])
+    })
   }
 
   [Symbol.for('nodejs.util.inspect.custom')] () {
@@ -353,7 +378,6 @@ module.exports = {
   fill,
   Headers,
   HeadersList,
-  binarySearch,
   normalizeAndValidateHeaderName,
   normalizeAndValidateHeaderValue
 }