--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -78,7 +78,7 @@ function fill (headers, object) {
 }
 
 class HeadersList {
-  constructor(init) {
+  constructor (init) {
     if (init instanceof HeadersList) {
       this[kHeadersMap] = new Map(init[kHeadersMap])
       this[kSorted] = init[kSorted]
@@ -117,7 +117,7 @@ class HeadersList {
     return this[kHeadersMap].has(normalizedName)
   }
 
-  sortMap() {
+  sortMap () {
     this[kHeadersMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
     this[kSorted] = true
   }
@@ -127,7 +127,7 @@ class HeadersList {
     return this[kHeadersMap].set(key, val)
   }
 
-  *[Symbol.iterator]() {
+  * [Symbol.iterator] () {
     if (!this[kSorted]) {
       this.sortMap()
     }
@@ -299,7 +299,7 @@ class Headers {
 
   // The value pairs to iterate over are the return value of running sort
   // and combine with this’s header list.
-  _getHeadersListIterable() {
+  _getHeadersListIterable () {
     if (!this[kHeadersList][kSorted]) {
       this[kHeadersList].sortMap()
     }