--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -12,7 +12,8 @@ const {
 } = require('./constants')
 
 const kHeadersMap = Symbol('headers map')
-const kSorted = Symbol('headers map sorted')
+const kSorted = Symbol('kSorted')
+const kHeadersSortedMap = Symbol('headers map sorted')
 
 function normalizeAndValidateHeaderName (name) {
   if (name === undefined) {
@@ -82,9 +83,11 @@ class HeadersList {
     if (init instanceof HeadersList) {
       this[kHeadersMap] = new Map(init[kHeadersMap])
       this[kSorted] = init[kSorted]
+      this[kHeadersSortedMap] = init[kHeadersSortedMap]
     } else {
       this[kHeadersMap] = new Map(init)
       this[kSorted] = false
+      this[kHeadersSortedMap] = new Map()
     }
   }
 
@@ -118,12 +121,13 @@ class HeadersList {
   }
 
   sortMap () {
-    this[kHeadersMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
+    this[kHeadersSortedMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
     this[kSorted] = true
   }
 
   // bypass normalization
   set (key, val) {
+    this[kSorted] = false
     return this[kHeadersMap].set(key, val)
   }
 
@@ -132,7 +136,7 @@ class HeadersList {
       this.sortMap()
     }
 
-    for (const [key, val] of this[kHeadersMap].entries()) {
+    for (const [key, val] of this[kHeadersSortedMap].entries()) {
       yield [key, val]
     }
   }
@@ -304,26 +308,26 @@ class Headers {
       this[kHeadersList].sortMap()
     }
 
-    return new HeadersList(this[kHeadersList])
+    return this[kHeadersList][kHeadersSortedMap]
   }
 
   * keys () {
-    const listFreezed = this._getHeadersListIterable()
-    for (const key of listFreezed[kHeadersMap].keys()) {
+    const sortedHeaders = this._getHeadersListIterable()
+    for (const key of sortedHeaders.keys()) {
       yield key
     }
   }
 
   * values () {
-    const listFreezed = this._getHeadersListIterable()
-    for (const val of listFreezed[kHeadersMap].values()) {
+    const sortedHeaders = this._getHeadersListIterable()
+    for (const val of sortedHeaders.values()) {
       yield val
     }
   }
 
   * entries () {
-    const listFreezed = this._getHeadersListIterable()
-    for (const [key, val] of listFreezed[kHeadersMap]) {
+    const sortedHeaders = this._getHeadersListIterable()
+    for (const [key, val] of sortedHeaders) {
       yield [key, val]
     }
   }
@@ -345,8 +349,8 @@ class Headers {
     const callback = args[0]
     const thisArg = args[1]
 
-    const listFreezed = this._getHeadersListIterable()
-    listFreezed[kHeadersMap].forEach((value, index) => {
+    const sortedHeaders = this._getHeadersListIterable()
+    sortedHeaders.forEach((value, index) => {
       callback.apply(thisArg, [value, index, this])
     })
   }