--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -12,7 +12,6 @@ const {
 } = require('./constants')
 
 const kHeadersMap = Symbol('headers map')
-const kSorted = Symbol('kSorted')
 const kHeadersSortedMap = Symbol('headers map sorted')
 
 function normalizeAndValidateHeaderName (name) {
@@ -82,12 +81,10 @@ class HeadersList {
   constructor (init) {
     if (init instanceof HeadersList) {
       this[kHeadersMap] = new Map(init[kHeadersMap])
-      this[kSorted] = init[kSorted]
       this[kHeadersSortedMap] = init[kHeadersSortedMap]
     } else {
       this[kHeadersMap] = new Map(init)
-      this[kSorted] = false
-      this[kHeadersSortedMap] = new Map()
+      this[kHeadersSortedMap] = undefined
     }
   }
 
@@ -101,7 +98,7 @@ class HeadersList {
       this[kHeadersMap].set(normalizedName, `${exists}, ${normalizedValue}`)
     } else {
       this[kHeadersMap].set(normalizedName, `${normalizedValue}`)
-      this[kSorted] = false
+      this[kHeadersSortedMap] = undefined
     }
   }
 
@@ -122,17 +119,16 @@ class HeadersList {
 
   sortMap () {
     this[kHeadersSortedMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
-    this[kSorted] = true
   }
 
   // bypass normalization
   set (key, val) {
-    this[kSorted] = false
+    this[kHeadersSortedMap] = undefined
     return this[kHeadersMap].set(key, val)
   }
 
   * [Symbol.iterator] () {
-    if (!this[kSorted]) {
+    if (!this[kHeadersSortedMap]) {
       this.sortMap()
     }
 
@@ -304,7 +300,7 @@ class Headers {
   // The value pairs to iterate over are the return value of running sort
   // and combine with this’s header list.
   _getHeadersListIterable () {
-    if (!this[kHeadersList][kSorted]) {
+    if (!this[kHeadersList][kHeadersSortedMap]) {
       this[kHeadersList].sortMap()
     }