--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -84,11 +84,13 @@ class HeadersList {
       this[kHeadersSortedMap] = init[kHeadersSortedMap]
     } else {
       this[kHeadersMap] = new Map(init)
-      this[kHeadersSortedMap] = undefined
+      this[kHeadersSortedMap] = null
     }
   }
 
   append (name, value) {
+    this[kHeadersSortedMap] = null
+
     const normalizedName = normalizeAndValidateHeaderName(name)
     const normalizedValue = normalizeAndValidateHeaderValue(name, value)
 
@@ -98,11 +100,19 @@ class HeadersList {
       this[kHeadersMap].set(normalizedName, `${exists}, ${normalizedValue}`)
     } else {
       this[kHeadersMap].set(normalizedName, `${normalizedValue}`)
-      this[kHeadersSortedMap] = undefined
     }
   }
 
+  set (name, value) {
+    this[kHeadersSortedMap] = null
+
+    const normalizedName = normalizeAndValidateHeaderName(name)
+    return this[kHeadersMap].set(normalizedName, value)
+  }
+
   delete (name) {
+    this[kHeadersSortedMap] = null
+
     const normalizedName = normalizeAndValidateHeaderName(name)
     return this[kHeadersMap].delete(normalizedName)
   }
@@ -117,24 +127,20 @@ class HeadersList {
     return this[kHeadersMap].has(normalizedName)
   }
 
-  sortMap () {
-    this[kHeadersSortedMap] = new Map([...this[kHeadersMap]].sort((a, b) => a[0] < b[0] ? -1 : 1))
+  keys() {
+    return this[kHeadersMap].keys()
   }
 
-  // bypass normalization
-  set (key, val) {
-    this[kHeadersSortedMap] = undefined
-    return this[kHeadersMap].set(key, val)
+  values () {
+    return this[kHeadersMap].values()
   }
 
-  * [Symbol.iterator] () {
-    if (!this[kHeadersSortedMap]) {
-      this.sortMap()
-    }
+  entries () {
+    return this[kHeadersMap].entries()
+  }
 
-    for (const [key, val] of this[kHeadersSortedMap].entries()) {
-      yield [key, val]
-    }
+  [Symbol.iterator] () {
+    return this[kHeadersMap][Symbol.iterator]()
   }
 }
 
@@ -276,20 +282,18 @@ class Headers {
       )
     }
 
-    const normalizedName = normalizeAndValidateHeaderName(String(args[0]))
-
     if (this[kGuard] === 'immutable') {
       throw new TypeError('immutable')
     } else if (
       this[kGuard] === 'request' &&
-      forbiddenHeaderNames.includes(normalizedName)
+      forbiddenHeaderNames.includes(String(args[0]).toLocaleLowerCase())
     ) {
       return
     } else if (this[kGuard] === 'request-no-cors') {
       // TODO
     } else if (
       this[kGuard] === 'response' &&
-      forbiddenResponseHeaderNames.includes(normalizedName)
+      forbiddenResponseHeaderNames.includes(String(args[0]).toLocaleLowerCase())
     ) {
       return
     }
@@ -297,35 +301,41 @@ class Headers {
     return this[kHeadersList].set(String(args[0]), String(args[1]))
   }
 
-  // The value pairs to iterate over are the return value of running sort
-  // and combine with this’s header list.
-  _getHeadersListIterable () {
-    if (!this[kHeadersList][kHeadersSortedMap]) {
-      this[kHeadersList].sortMap()
-    }
-
+  get [kHeadersSortedMap] () {
+    this[kHeadersList][kHeadersSortedMap] ??= new Map([...this[kHeadersList]].sort((a, b) => a[0] < b[0] ? -1 : 1))
     return this[kHeadersList][kHeadersSortedMap]
   }
 
-  * keys () {
-    const sortedHeaders = this._getHeadersListIterable()
-    for (const key of sortedHeaders.keys()) {
-      yield key
+  keys () {
+    if (!(this instanceof Headers)) {
+      throw new TypeError('Illegal invocation')
     }
+
+    return this[kHeadersSortedMap].keys()
   }
 
-  * values () {
-    const sortedHeaders = this._getHeadersListIterable()
-    for (const val of sortedHeaders.values()) {
-      yield val
+  values () {
+    if (!(this instanceof Headers)) {
+      throw new TypeError('Illegal invocation')
     }
+
+    return this[kHeadersSortedMap].values()
   }
 
-  * entries () {
-    const sortedHeaders = this._getHeadersListIterable()
-    for (const [key, val] of sortedHeaders) {
-      yield [key, val]
+  entries () {
+    if (!(this instanceof Headers)) {
+      throw new TypeError('Illegal invocation')
     }
+
+    return this[kHeadersSortedMap].entries()
+  }
+
+  [Symbol.iterator] () {
+    if (!(this instanceof Headers)) {
+      throw new TypeError('Illegal invocation')
+    }
+
+    return this[kHeadersSortedMap]
   }
 
   forEach (...args) {
@@ -345,8 +355,7 @@ class Headers {
     const callback = args[0]
     const thisArg = args[1]
 
-    const sortedHeaders = this._getHeadersListIterable()
-    sortedHeaders.forEach((value, index) => {
+    this[kHeadersSortedMap].forEach((value, index) => {
       callback.apply(thisArg, [value, index, this])
     })
   }