--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -127,7 +127,7 @@ class HeadersList {
     return this[kHeadersMap].has(normalizedName)
   }
 
-  keys() {
+  keys () {
     return this[kHeadersMap].keys()
   }