--- a/lib/fetch/index.js
+++ b/lib/fetch/index.js
@@ -1334,7 +1334,7 @@ async function httpNetworkOrCacheFetch (
   //    user agents should append `User-Agent`/default `User-Agent` value to
   //    httpRequest’s header list.
   if (!httpRequest.headersList.has('user-agent')) {
-    httpRequest.headersList.append('user-agent', 'undici')
+    httpRequest.headersList.set('user-agent', 'undici')
   }
 
   //    15. If httpRequest’s cache mode is "default" and httpRequest’s header
@@ -1361,7 +1361,7 @@ async function httpNetworkOrCacheFetch (
     !httpRequest.preventNoCacheCacheControlHeaderModification &&
     !httpRequest.headersList.has('cache-control')
   ) {
-    httpRequest.headersList.append('cache-control', 'max-age=0')
+    httpRequest.headersList.set('cache-control', 'max-age=0')
   }
 
   //    17. If httpRequest’s cache mode is "no-store" or "reload", then:
@@ -1369,20 +1369,20 @@ async function httpNetworkOrCacheFetch (
     // 1. If httpRequest’s header list does not contain `Pragma`, then append
     // `Pragma`/`no-cache` to httpRequest’s header list.
     if (!httpRequest.headersList.has('pragma')) {
-      httpRequest.headersList.append('pragma', 'no-cache')
+      httpRequest.headersList.set('pragma', 'no-cache')
     }
 
     // 2. If httpRequest’s header list does not contain `Cache-Control`,
     // then append `Cache-Control`/`no-cache` to httpRequest’s header list.
     if (!httpRequest.headersList.has('cache-control')) {
-      httpRequest.headersList.append('cache-control', 'no-cache')
+      httpRequest.headersList.set('cache-control', 'no-cache')
     }
   }
 
   //    18. If httpRequest’s header list contains `Range`, then append
   //    `Accept-Encoding`/`identity` to httpRequest’s header list.
   if (httpRequest.headersList.has('range')) {
-    httpRequest.headersList.append('accept-encoding', 'identity')
+    httpRequest.headersList.set('accept-encoding', 'identity')
   }
 
   //    19. Modify httpRequest’s header list per HTTP. Do not append a given
@@ -1390,9 +1390,9 @@ async function httpNetworkOrCacheFetch (
   //    TODO: https://github.com/whatwg/fetch/issues/1285#issuecomment-896560129
   if (!httpRequest.headersList.has('accept-encoding')) {
     if (/^https:/.test(requestCurrentURL(httpRequest).protocol)) {
-      httpRequest.headersList.append('accept-encoding', 'br, gzip, deflate')
+      httpRequest.headersList.set('accept-encoding', 'br, gzip, deflate')
     } else {
-      httpRequest.headersList.append('accept-encoding', 'gzip, deflate')
+      httpRequest.headersList.set('accept-encoding', 'gzip, deflate')
     }
   }
 
@@ -1918,7 +1918,7 @@ async function httpNetworkFetch (
         origin: url.origin,
         method: request.method,
         body: fetchParams.controller.dispatcher[kIsMockActive] ? request.body && request.body.source : body,
-        headers: request.headersList,
+        headers: [...request.headersList].flat(),
         maxRedirections: 0
       },
       {