--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -464,6 +464,7 @@ class Request {
       // this’s headers.
       if (contentType && !this[kHeaders].has('content-type')) {
         this[kHeaders].append('content-type', contentType)
+        this[kState].headersList.append('content-type', contentType)
       }
     }