… lib/fetch/response.js
… test/fetch/headers.js
… test/node-fetch/mock.js