This PR replaces the `HeadersList` Array inheritance to Map, which basically is more suitable for headers use case. ~It also increases the performance by 10% (it's not done yet, so it might be more/less)~.