@Ethan-Arrowood 2022-03-29
From my investigations the Map wasn't as good for reading the header values (as in per spec you must return the values in sorted order) but I do know v8 is always making magic happen; especially with objects like Map. 

Wouldn't be shocked if your new benchmark results are even better perf wise. 

One thing to keep in mind is memory footprint too; a flat array of strings should be more efficient space wise but it's a hard one to verify as well.