@ronag 2022-03-29
> > as in per spec you must return the values in sorted order
> 
> Can you point me to this statement in the spec? Yeah, in my benchmarks it was better than Array implementation, mostly due to the fact of needing `binarySearch` to do any operation in headers using `Array` inheritance

https://fetch.spec.whatwg.org/#headers-class

> The [value pairs to iterate over](https://webidl.spec.whatwg.org/#dfn-value-pairs-to-iterate-over) are the return value of running [sort and combine](https://fetch.spec.whatwg.org/#concept-header-list-sort-and-combine) with [this](https://webidl.spec.whatwg.org/#this)’s [header list](https://fetch.spec.whatwg.org/#concept-headers-header-list).