@AlttiRi 2022-03-29
Just run a server that logs the incoming request's headers:

![Chrome headers](https://user-images.githubusercontent.com/16310547/160640902-5debb336-1c76-4f08-a486-00b2538fdae1.png)

The headers are **not** alphabetically ordered. 
Neither in requests from `fetch` function, nor in any other request from the browser.

Also I don't see any clear mention in the specs that the headers should be alphabetically ordered.

BTW, an example of _"sorted"_ array: `["1", "2", "3", "4", "5"].sort(() => Math.random() >= 0.5 ? -1 : 1)`.

---

I expect that it (undici/fetch) will allow to a user decide the order of the headers, if the user will decide to overwrite the default headers, or add new ones.