@RafaelGSS 2022-03-30
In fact, the iteration is sorted: 

![image](https://user-images.githubusercontent.com/26234614/160926103-02760509-80cf-4dc1-bd03-9bfdf4c80c3a.png)

I'll adjust the PR to match this requirement.