@RafaelGSS 2022-03-31
I have adjusted, but still non-optimal. I'll implement a orderedmap soon.