@ronag 2022-04-09
@RafaelGSS I helped a little. Hope you don't mind. PTAL.